这2个chat单纯是我run的时候出error而已，PostJobScreen是刚开始的Date和time点了没反应，改成有反应。

+++ app/src/main/java/com/example/microjob/viewmodel/ChatViewModel.kt
@@ -275,7 +275,7 @@ class ChatViewModel(
         workerRating: Float = 0f,
         workerComment: String = "",
         onReleased: () -> Unit,
-        onJobCompleted: (jobTitle: String, jobPrice: Int) -> Unit = { _, _ -> }
+        onJobCompleted: (jobTitle: String, jobPrice: Double) -> Unit = { _, _ -> }
     ) {
         viewModelScope.launch {
             try {
